extract all availability messaging
