+++ src/components/ItemTable/ItemAvailabilityComponents/NotAvailable.tsx
@@ -1,0 +1,17 @@
+import { Box } from "@nypl/design-system-react-components"
+import type Item from "../../../models/Item"
+import ContactALibrarian from "./ContactALibrarian"
+
+const NotAvailable = ({ item }: { item: Item }) => {
+  return (
+    <>
+      <Box as="span" color="ui.warning.tertiary">
+        Not available
+      </Box>
+      {item.dueDate && ` - In use until ${item.dueDate}`}
+      <ContactALibrarian item={item} />
+    </>
+  )
+}
+
+export default NotAvailable
